Swift 3 support (#65)
* Update to Swift 3

* Bump version 0.4.0

* Rename HTTPURLResponse file too

* Prefer private over fileprivate

* Fix serializating HTTPURLResponse

* Don't say we failed if we didn't

* Swift 3.0.1

* tvOS target

* Swift 3.0.1

* Fix tvOS deployment target

* Fix upload tests

* Explicitly set SWIFT_VERSION

Without this, Xcode complains about the legacy Swift version not being
set.

* Remove explicit toolchain from CI

* Specify SDK for CI

* Add destinations for CI

...because Travis CI is so good.

* Use latest version of Xcode

* Clean up spacing

+++ DVR/Interaction.swift
@@ -93,7 +93,7 @@ extension Interaction {
     init?(dictionary: [String: Any]) {
         guard let request = dictionary["request"] as? [String: Any],
             let response = dictionary["response"] as? [String: Any],
-            let recordedAt = dictionary["recorded_at"] as? Double else { return nil }
+            let recordedAt = dictionary["recorded_at"] as? Int else { return nil }
 
         self.request = NSMutableURLRequest(dictionary: request) as URLRequest
         self.response = HTTPURLResponse(dictionary: response)
